Trying to make a small platformer creator
I don't know if I'll mimick THAT platformer maker at 100%, but
I'd like to add some features available there.

Still, at the moment, it's just a TileMap that can be filled with
"blocks" terrain tiles, and which export the whole thing as an
incomprehensible JSON array.

Still, the JSON array was parsed successfully on VRChat, to recreate
the created "level" in 3D.

Signed-off-by: Voyage <voyage@miouyouyou.fr>